Rework conftest.
